+++ src/templates/states.rs
@@ -9,19 +9,14 @@ pub use utils::*;"#
 
 pub fn state_rs() -> &'static str {
     r#"use super::utils::{load_acc_mut_unchecked, DataLen};
-use pinocchio::{
-    account_info::AccountInfo,
-    program_error::ProgramError,
-    pubkey::{self, Pubkey},
-    ProgramResult,
-};
+use pinocchio::{error::ProgramError, Address, AccountView, ProgramResult};
 
 use crate::{errors::MyProgramError, instructions::Initialize};
 
 #[repr(C)]
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub struct MyState {
-    pub owner: Pubkey,
+    pub owner: Address,
 }
 
 impl DataLen for MyState {
@@ -31,18 +26,18 @@ impl DataLen for MyState {
 impl MyState {
     pub const SEED: &'static str = "init";
 
-    pub fn validate_pda(bump: u8, pda: &Pubkey, owner: &Pubkey) -> Result<(), ProgramError> {
-        let seed_with_bump = &[Self::SEED.as_bytes(), owner, &[bump]];
-        let derived = pubkey::create_program_address(seed_with_bump, &crate::ID)?;
+    pub fn validate_pda(bump: u8, pda: &Address, owner: &Address) -> Result<(), ProgramError> {
+        let seed_with_bump = &[Self::SEED.as_bytes(), owner.as_array().as_slice(), &[bump]];
+        let derived = Address::create_program_address(seed_with_bump, &crate::ID)?;
         if derived != *pda {
             return Err(MyProgramError::PdaMismatch.into());
         }
         Ok(())
     }
 
-    pub fn initialize(my_stata_acc: &AccountInfo, ix_data: &Initialize) -> ProgramResult {
+    pub fn initialize(my_stata_acc: &mut AccountView, ix_data: &Initialize) -> ProgramResult {
         let my_state =
-            unsafe { load_acc_mut_unchecked::<MyState>(my_stata_acc.borrow_mut_data_unchecked()) }?;
+            unsafe { load_acc_mut_unchecked::<MyState>(my_stata_acc.borrow_unchecked_mut()) }?;
 
         my_state.owner = ix_data.owner;
         Ok(())
@@ -51,7 +46,7 @@ impl MyState {
 }
 
 pub fn utils_rs() -> &'static str {
-    r#"use pinocchio::program_error::ProgramError;
+    r#"use pinocchio::error::ProgramError;
 
 use crate::errors::MyProgramError;